Small cleanup of self tests
